Please enter the commit message for your changes. Lines starting  with '' will be ignored, and an empty message aborts the commit.
 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   app/Http/Controllers/Admin/MovieController.php
	modified:   app/Http/Controllers/IndexController.php
	modified:   app/Models/Category.php
	new file:   database/migrations/2024_03_21_171812_add_slug_column_to_movies_table.php
	new file:   public/uploads/movie/423686686_1571427743675019_8256050145316757786_n-8541.jpg
	deleted:    public/uploads/movie/Screenshot 2024-03-09 011021-171.png
	new file:   public/uploads/movie/image-6438.png
	new file:   public/uploads/movie/image-929.png
	renamed:    public/uploads/movie/product-01-9105.jpg -> public/uploads/movie/product-01-1746.jpg
	new file:   public/uploads/movie/product-01-2349.jpg
	deleted:    public/uploads/movie/product-07-1304.jpg
	new file:   public/uploads/movie/product-15-7610.jpg
	modified:   resources/views/pages/category.blade.php
	modified:   resources/views/pages/country.blade.php
	modified:   resources/views/pages/genre.blade.php
	modified:   resources/views/pages/home.blade.php
	modified:   resources/views/pages/movie.blade.php
	modified:   routes/web.php

+++ app/Http/Controllers/Admin/MovieController.php
@@ -66,11 +66,11 @@ class MovieController extends Controller
                 'title'=>$request->title,
                 'description'=>$request->description,
                 'status' => $request->status,
-                // 'slug' => $request->slug,
+                'slug' => $request->slug,
                 'image'=> $image['image'],
                 'country_id' => $request->country_id,
                 'name_eng' => $request->name_eng,
-
+                
                 'genre_id' => $request->genre_id,
                 'hot_movie' => $request->hot_movie,
 
@@ -131,7 +131,7 @@ class MovieController extends Controller
         $image = $this->movieService->uploadImage($request);
         $movie = Movie::find($id);
         $get_image = $request->file('image');
-        if(!empty($movie->image)){
+        if(!empty($movie->image) && isset($request->image)){
             unlink('uploads/movie/'.$movie->image);
         }
         if($image != false){
@@ -141,8 +141,7 @@ class MovieController extends Controller
                 'status' => $request->status,
                 'slug' => $request->slug,
                 'name_eng' => $request->name_eng,
-
-                'image'=> $image['image'],
+                'image'=> $image['image'] ?? null,
                 'country_id' => $request->country_id,
                 'genre_id' => $request->genre_id,
                 'category_id' => $request->category_id,
@@ -151,7 +150,7 @@ class MovieController extends Controller
             ]);
             return redirect()->back();
         }
-        return false;
+        // return false;
     }
 
     /**
